Standardize letter case and trim whitespace for dataframes

Removes leading and trailing whitespace from every column in a dataframe or list of dataframes. At the same time it standardizes the case of the characters to either upper or lower case. This function works best with raw character data before converting data types and other data munging/feature engineering. Does handle a dataframe as it is, however, and returns every column in it's original data type as a dataframe.

casetrim(df, case = c("upper", "lower"), colclasses = c("asis",
  "character"))

Arguments

df
a dataframe or list of dataframes with columns of any character type

case
the letter case you want the output to be: upper or lower. Default is upper.

colclasses
do you want to return every column to it's original data type (asis) or simply return all cols as characters (character)? Default is asis.

See also

Examples

df <- data.frame(
    performingPhysio = c("JILL jones      ", "    jack BLACK", "", " RegiNald ChesterField ")
  , operatingPhysio  = c("Who Dat      ", "chuck FLEMMING      ", NA, "          shankman, razor shankman")
  , consent          = rep(FALSE, times = 4)
  , duration_hrs     = 5:8
  , stringsAsFactors = FALSE
  )

casetrim(df)

#>        performingphysio          operatingphysio consent duration_hrs
#> 1            JILL JONES                  WHO DAT   FALSE            5
#> 2            JACK BLACK           CHUCK FLEMMING   FALSE            6
#> 3                                           <NA>   FALSE            7
#> 4 REGINALD CHESTERFIELD SHANKMAN, RAZOR SHANKMAN   FALSE            8

casetrim(
    df
  , case = 'lower'
)

#>        performingphysio          operatingphysio consent duration_hrs
#> 1            jill jones                  who dat   FALSE            5
#> 2            jack black           chuck flemming   FALSE            6
#> 3                                           <NA>   FALSE            7
#> 4 reginald chesterfield shankman, razor shankman   FALSE            8

casetrim(
    df
  , case       = 'lower'
  , colclasses = 'character'
)

#>        performingphysio          operatingphysio consent duration_hrs
#> 1            jill jones                  who dat   false            5
#> 2            jack black           chuck flemming   false            6
#> 3                                           <NA>   false            7
#> 4 reginald chesterfield shankman, razor shankman   false            8