Wrapper for paste and print

Takes a vector of objects and puts them together in a vector of strings. It's useful to track progress of for loops and to automate output.

pp(...)

Arguments

...
A collection of arguments for paste().

See also

Examples

pp(letters[1:10], 11:20)

#>  [1] "a 11" "b 12" "c 13" "d 14" "e 15" "f 16" "g 17" "h 18" "i 19" "j 20"

for(i in 1:5){pp("This is the capital letter ", toupper(letters[i]), ".", sep = "")}

#> [1] "This is the capital letter A."
#> [1] "This is the capital letter B."
#> [1] "This is the capital letter C."
#> [1] "This is the capital letter D."
#> [1] "This is the capital letter E."