Convert string to camelcase (ie PaulJames)

Takes a string, vector(s) of strings, or list of strings and converts the case to camelcase. Outputs a string vector if only a single string vector was used as input. Outputs a list of string vectors if multiple string vectors were used as input. Outputs a single string if the collapseall flag is set to TRUE.

tocamel(..., sep = "\\W+", collapseall = FALSE)

Arguments

...
A string, vector of strings, or list of strings in any letter case pattern.

sep
A character vector containing characters or regular expression(s) to use for splitting. Default is \\W+ (one or more characters that are not word characters; IE: not a letter or number).

collapseall
A logical/boolean flag to indicated whether or not you want a single string as your output (TRUE), or multiple strings (FALSE). Default is FALSE.

See also

Examples

tocamel("jonny appleseed")

#> [1] "JonnyAppleseed"

tocamel(c("THE", "DARK", "KNIGHT"), c("rises from the lazerus PIT"))

#> [[1]]
#> [1] "The"    "Dark"   "Knight"
#> 
#> [[2]]
#> [1] "RisesFromTheLazerusPit"
#> 

tocamel(c("THE", "DARK", "KNIGHT"), c("rises from the lazerus PIT"), collapseall = TRUE)

#> [1] "TheDarkKnightRisesFromTheLazerusPit"

tocamel("GraNDpa CrIPEs -mcgee!!!!")

#> [1] "GrandpaCripesMcgee"