Chainable View

A View that doesn't break a pipeline. Same thing as utils::View(), but returns the original data object invisibly to allow putting view commands anywhere in a pipeline. This also allows for adding a view command at the end of pipelines while also assigning the data object the pipeline returns.

view(x, title)

Arguments

x
an R object which can be coerced to a data frame with non-zero numbers of rows and columns

title
the title for the viewer window; defaults to name of x if not given

See also

Examples

# NOT RUN {
# 1
view(iris)

# 2
iris_test <- iris %>%
  group_by(
    Species
  ) %>%
  summarise(
    sum = sum(Petal.Width)
  ) %>%
  view('iris:sum') %>%
  mutate(
    sum_squared = sum^2
  ) %>%
  view('iris:sum_squared')
# }
#